add file structure, refactor code
