+++ index.html
@@ -24,6 +24,6 @@
 <footer class="footer">
     <p class="footer__text">With 🖤️ by <a href="https://github.com/raveoverdose">RaveOverdose</a></p>
 </footer>
-<script src="./script/app.js"></script>
+<script type="module" src="./script/app.js"></script>
 </body>
 </html>
