+++ script/config.js
@@ -1,0 +1,3 @@
+export const API_KEY_OW = '&appid=0bed307b0afbac425fa1b0935e9651de';
+export const API_KEY_IP = 'at_eRqcqqInNt1bUk30MoByCCUV19sX3';
+export const container = document.querySelector('.container');
